# admin/src/main/java/com/github/nk/klusterfuck/admin/controllers/objects/RefreshRequest.java
package com.github.nk.klusterfuck.admin.controllers.objects;

/**
 * Created by nk on 20/6/17.
 */
public class RefreshRequest {
	private String refreshToken;

	public String getRefreshToken() {
		return refreshToken;
	}

	public void setRefreshToken(String refreshToken) {
		this.refreshToken = refreshToken;
	}
}