# client/src/main/resources/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "API",
    "version" : "v1",
    "title" : "API"
  },
  "basePath" : "/",
...
  "schemes" : [ "http" ],
  "paths" : {
    "/api/v1/auth/login" : {
      "post" : {
        "tags" : [ "Auth" ],
        "summary" : "login",
        "description" : "",
        "operationId" : "login",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/LoginRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/AccesstokenResponseWrapper"
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh" : {
      "post" : {
        "tags" : [ "Auth" ],
        "summary" : "refresh",
        "description" : "",
        "operationId" : "refresh",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/RefreshRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/AccesstokenResponseWrapper"
            }
          }
        }
      }
    },
    "/api/v1/auth/register" : {
      "post" : {
        "tags" : [ "Auth" ],
        "summary" : "register",
        "description" : "",
        "operationId" : "register",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/RegisterRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/api/v1/flows" : {
      "get" : {
        "tags" : [ "Flows" ],
        "summary" : "list",
        "description" : "",
        "operationId" : "list",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Flow"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Flows" ],
        "summary" : "create",
        "description" : "",
        "operationId" : "create",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/CreateFlowRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Flow"
            }
          }
        }
      }
    },
    "/api/v1/flows/validate" : {
      "post" : {
        "tags" : [ "Flows" ],
        "summary" : "validate",
        "description" : "",
        "operationId" : "validate",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
    "/api/v1/flows/{id}" : {
      "get" : {
        "tags" : [ "Flows" ],
        "summary" : "get",
        "description" : "",
        "operationId" : "get",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Flow"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Flows" ],
        "summary" : "delete",
        "description" : "",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
    "/api/v1/flows/{id}/deploy" : {
      "post" : {
        "tags" : [ "Flows" ],
        "summary" : "deploy",
        "description" : "",
        "operationId" : "deploy",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Flow"
            }
          }
        }
      }
    },
    "/api/v1/flows/{id}/model" : {
      "get" : {
        "tags" : [ "Flows" ],
        "summary" : "getModel",
        "description" : "",
        "operationId" : "getModel",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/DAGStepRef"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Flows" ],
        "summary" : "saveModel",
        "description" : "",
        "operationId" : "saveModel",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
    "AccesstokenResponseWrapper" : {
      "type" : "object",
      "properties" : {
        "sessionState" : {
          "type" : "string"
        },
        "token" : {
          "type" : "string"
        },
        "refreshToken" : {
          "type" : "string"
        },
        "expiresIn" : {
          "type" : "integer",
          "format" : "int64"
        },
        "refreshExpiresIn" : {
          "type" : "integer",
          "format" : "int64"
        },
        "tokenType" : {
          "type" : "string"
        },
        "idToken" : {
          "type" : "string"
        },
        "notBeforePolicy" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
    "CreateFlowRequest" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string",
          "enum" : [ "generic", "nodejs", "static_site" ]
        }
      }
    },
...
    "DAGStepRef" : {
      "type" : "object",
      "properties" : {
        "nodes" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/NodeStepRef"
          }
        },
        "links" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Link"
          }
        }
      }
    },
    "Flow" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "owner" : {
          "$ref" : "#/definitions/UserNamespace"
        },
        "name" : {
          "type" : "string"
        },
        "displayName" : {
          "type" : "string"
        }
      }
    },
...
    "Link" : {
      "type" : "object",
      "properties" : {
        "from" : {
          "type" : "string"
        },
        "to" : {
          "type" : "string"
        },
        "uiProps" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        }
      }
    },
...
    "LoginRequest" : {
      "type" : "object",
      "properties" : {
        "username" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        }
      }
    },
    "NodeStepRef" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "data" : {
          "$ref" : "#/definitions/StepRef"
        },
        "uiProps" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        }
      }
    },
...
    "RefreshRequest" : {
      "type" : "object",
      "properties" : {
        "refreshToken" : {
          "type" : "string"
        }
      }
    },
    "RegisterRequest" : {
      "type" : "object",
      "required" : [ "email", "firstName", "lastName", "password", "username" ],
      "properties" : {
        "firstName" : {
          "type" : "string"
        },
        "lastName" : {
          "type" : "string"
        },
        "username" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        }
      }
    },
...
    "StepRef" : {
      "type" : "object",
      "discriminator" : "category",
      "properties" : {
        "category" : {
          "type" : "string",
          "enum" : [ "fn", "connector" ]
        },
        "text" : {
          "type" : "string"
        }
      }
    },
...
    "UserNamespace" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "displayName" : {
          "type" : "string"
        },
        "gitUser" : {
          "type" : "string"
        },
        "gitPassword" : {
          "type" : "string"
        }
      }
    },
...
  }
}